moved paginate to its own file
